+++ CHANGELOG.md
@@ -8,11 +8,13 @@ All notable changes to this project will be documented in this file.
 ### Added
 
 - Added `default_schema_id` field to the `schema_registry_decode` processor. (@mmatczuk)
+- Go API: Component linter added to `public/schema`, including Redpanda build meta fields. (@Jeffail)
 
 ### Changed
 
 - (google_cloud_storage) Field `bucket` can now be interpolated (@rockwotj)
 - (output_sns) Field `topic_arn` can now be interpolation (@josephwoodward)
+- (Benthos) Logging: Enable timestamp output by default (@josephwoodward)
 
 ## 4.63.0 - 2025-08-27